requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
setuptools==75.6.0
sqlparse==0.5.3
tzdata==2024.2
wheel==0.45.1
